backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
google-genai>=0.7.0
sqlmodel>=0.0.16
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
